+++ .gitignore
@@ -14,3 +14,5 @@ __pycache__/*
 *.ipynb
 freesound_api.json
 queries.json
+
+*.pyc
